==> 1-Pokemon_Table_formatting.ipynb <==
# Cleaning & Table Formatting

---

In this notebook I will be cleaning and editing and formating some data used in the rest of the analysis

import pandas as pd
poke = pd.read_csv('./Data/pokedex_formal.csv')

# Creating a common type combo for coloring purposes.  
poke.head()

# Issues when i insitially read this in that needed to be cleaned up
def type_clean(types):
    chars = chars = ["'",'[',']',' ']
    cleaned_type = ''.join(letter for letter in types if letter not in chars)
    try:
        cleaned_list = cleaned_type.split(',')
        return str(sorted(cleaned_list))
    except:
        return cleaned_type

poke['Type'] = poke['Type'].apply(type_clean)
poke.head()

# Value counts of current typings and type-sets
poke['Type'].value_counts()


poke['Type_P'].value_counts()+poke['Type_S'].value_counts()

type_count_df = pd.concat([pd.DataFrame(poke['Type_P'].value_counts()),pd.DataFrame(poke['Type_S'].value_counts())],
          axis=1)
type_count_df['Total'] = type_count_df['Type_P']+type_count_df['Type_S'] 
type_count_df.reset_index(inplace=True)

#!pip install pandasql
import pandasql
from pandasql import sqldf

type_sum = pd.DataFrame(columns = ['Type','Total','HP','Attack','Defense','SpecialAttack','SpecialDefense','Speed','Count'])
type_sum

# Using a sql qeury to create sums and averages of individual types 
# 
for a_type in type_count_df['index']:
    query = '''
SELECT Name, Total, HP, Attack, Defense, SpecialAttack, SpecialDefense, Speed
FROM poke
WHERE Type_P ="{}" OR Type_S ="{}";
'''.format(a_type, a_type)
    a_type_df = sqldf(query, globals())
    a_type_mean = a_type_df.describe().loc['mean']
    
    type_sum.loc[len(type_sum)]=[a_type, a_type_mean[0],a_type_mean[1],a_type_mean[2],a_type_mean[3],
                               a_type_mean[4],a_type_mean[5],a_type_mean[6],int(a_type_df.describe().loc['count'][0])]

int(a_type_df.describe().loc['count'][0])

type_sum

type_summary_df = pd.merge(left = type_count_df, right= type_sum, left_on = 'index', right_on = 'Type' )
type_summary_df.rename(columns = {'Total_x':'Occurances'}, inplace=True)

type_summary_df.rename(columns = {'index':'Type','Type_P':'Primary_Type_Occ','Type_S':'Secondary_Type_Occ','Occurances':'Total_Occ' }, inplace=True)
type_summary_df.to_csv('./Data/Pokemon_Summary.csv')

subtype_df.describe().loc['mean']
