==> 2-Poke-Table-Creation.ipynb <==
# Table Creation

Going to Create a pivot table showing type-by-type counts as well as type-by-type attacking & defensive advantages/disadvantages.

The advantages table was taken from Git Hub User [Zonination](https://github.com/zonination/).  Thanks you!


import pandas as pd


poke = pd.read_csv('pokedex_formal_2.csv')
poke.head(5)

#### For the sake of this analysis I will not be considering Mega Evolutions as they are more or less an empowered state.  I'll also be examining Legendary(Mythical) Pokemon separately.  

poke = poke[poke['Mega'] ==0]
poke = poke[poke['legendary'] ==0]

types = sorted(list(set(list(poke.Type_P) +list(poke.Type_S))))
types

import itertools
poke_pivot = pd.DataFrame(index = types, columns = types)
poke_pivot.fillna(value = 0, inplace = True)

for TP, TS in zip(poke['Type_S'].values, poke['Type_P'].values ):

    for type_combo in itertools.combinations_with_replacement(types, 2):
        
        if type_combo[0] == TP and type_combo[1]== TS:
            poke_pivot.at[type_combo[1], type_combo[0]] +=1
            if type_combo[0] != type_combo[1]:
                poke_pivot.at[type_combo[0], type_combo[1]] +=1


            
        if type_combo[1] == TP and type_combo[0]== TS:
            poke_pivot.at[type_combo[1], type_combo[0]] +=1
            if type_combo[0] != type_combo[1]:
                poke_pivot.at[type_combo[0], type_combo[1]] +=1

    
poke_pivot

poke_pivot['Total'] = poke_pivot.sum(axis =1)

poke_pivot

poke_pivot.to_csv('poke_pivot.csv')

# Thanks GitHub user zonination for the base version.
advantages = pd.read_csv('https://raw.githubusercontent.com/zonination/pokemon-chart/master/chart.csv', 
                         index_col =0)
#advantages
advantages = advantages.reindex_axis(poke_pivot.columns[:-1], axis =1)
advantages = advantages.reindex_axis(poke_pivot.columns[:-1], axis =0)

advantages.to_csv('advantages.csv')

advantages
